--- src/lib/core/index.ts
@@ -1,1 +1,0 @@
-export * as dialog from './dialog/index.js';
