+++ src/lib/index.ts
@@ -1,4 +1,4 @@
-export * from './core/index.js';
+export * as syv from './syv.js';
 export { default as ButtonLink } from './core/ButtonLink.svelte';
 export { default as Dialog } from './core/Dialog.svelte';
 export { default as GradientBorder } from './core/GradientBorder.svelte';
